backend/requirements/requirements.txt
# Core
# ------------------------------------------------------------------------------
pytz==2023.3  # https://github.com/stub42/pytz
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
python-decouple==3.8 # https://github.com/henriquebastos/python-decouple
packaging==23.1 # https://pypi.org/project/packaging/
psycopg2==2.9.6 # https://pypi.org/project/psycopg2/

# Django
# ------------------------------------------------------------------------------
django==5.2.4  # pyup: < 3.1  # https://www.djangoproject.com/
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-filter==23.2  # https://github.com/carltongibson/django-filter
django-phonenumbers==1.0.1
django-phonenumber-field==7.1.0
django-cors-headers==4.3.1

# Django REST Framework
# ------------------------------------------------------------------------------
djangorestframework==3.16.0  # https://github.com/encode/django-rest-framework
drf-writable-nested==0.7.0  # https://github.com/beda-software/drf-writable-nested
drf-spectacular==0.26.0  # https://github.com/tfranzel/drf-spectacular
drf-spectacular-sidecar==2023.3.1

# Celery
# ------------------------------------------------------------------------------
celery==5.3.1  # https://github.com/celery/celery
kombu==5.3.1  # https://github.com/celery/kombu
flower==1.0.0  # https://github.com/mher/flower
redis==4.5.5 # https://github.com/andymccurdy/redis-py
django-redis==5.3.0

# Other
# ------------------------------------------------------------------------------
Unidecode==1.4.0